[clang-tidy] General fixes
- Silenced narrowing warnings [cppcoreguidelines-narrowing-conversions]
- Silenced cppcoreguidelines-macro-usage because macro is needed to provide `__FUNCTION__`,
  `__FILE__`, `__LINE__` to the invoked function, and simplified to improve readability by not
  having a double negative (`#ifndef NDEBUG`). And only define `assert_x()` when `NDEBUG` isn't
  defined
- Silenced cppcoreguidelines-pro-type-reinterpret-cast due to how `Q_DECLARE_METATYPE()` works
  internally

+++ src/Util.cc
@@ -214,8 +214,8 @@ void Util::resizeRatioOfScreen(QWidget *widget, float percentage, const QSize &m
   if (screen == nullptr) return;
 
   auto rect = screen->availableGeometry();
-  rect.setWidth(static_cast<float>(rect.width()) * percentage);
-  rect.setHeight(static_cast<float>(rect.height()) * percentage);
+  rect.setWidth(int(float(rect.width()) * percentage));
+  rect.setHeight(int(float(rect.height()) * percentage));
 
   auto size = rect.size();
   if (size.width() < minimum.width()) {
